refactor code for choiceinput
