+++ Composer/packages/extensions/visual-designer/src/components/nodes/index.tsx
@@ -11,3 +11,4 @@ export * from './ReplaceDialog';
 export * from './SwitchCondition';
 export * from './UnknownIntentRule';
 export * from './ConversationUpdateActivityRule';
+export * from './ChoiceInput';
